requirements.txt
streamlit>=1.20.0
Pillow>=9.0.0
numpy>=1.20.0
pandas>=1.3.0
fpdf>=1.7.2
